requirements.txt
langchain
langchain-community
langchain-ollama
langchain_openai
openai
gradio
requests
beautifulsoup4
ddgs
python-dotenv
typing-extensions
